make_appointment/requirements.txt
Django==1.8.7
django-bootstrap3-datetimepicker==2.2.3
flake8==2.5.1
mccabe==0.3.1
pep8==1.5.7
pyflakes==1.0.0
wheel==0.24.0
